moved the alias matching functions to new module name to better fit in with the crowd
